fix(TESB-22183): WSDLLoader refinement for multiple imports for a namespace. (#1694)
* fix(TESB-22183): Add support for circular WSDL imports to WSDLLoader.

Circular WSDL imports are not prohibited by WS-I basic profile. However, they are currently causing WSDL import to fail in an endless recursion. The present fix corrects this condition and ensures that WSDL documents are not imported more than once.

* fix(TESB-22183): Attribute name error in fix corrected.

* fix(TESB-22183): "location" attribute corrected for circular imports.

* fix(TESB-22183): Fix support for circular WSDL imports to WSDLLoader.

Imports causing circular WSDL document dependencies are now removed, because WSDL4J does not handle them properly, but runs into an endless recursion, but processes the WSDL as expected when the "backwards" import is removed.

* fix(TESB-22183): Add support for circular WSDL imports to WSDLLoader.

Circular WSDL imports are not prohibited by WS-I basic profile. However, they are currently causing WSDL import to fail in an endless recursion. The present fix corrects this condition and ensures that WSDL documents are not imported more than once.