fix(TUP-22906) change follow ref also to find merge
Previously, we only follow the main link to find the first merge
component.
But if there is any reference link, it will not be able to find then the
first component to start.
(tFileInputDelimited_5 here normally was not able to find the first
merge from XML component, then find back the design node
tFileInputDelimited_4)